Add files via upload
Add CollectionView to display a list of films with their names, images, and prices. Implement FilmCell to customize the appearance of each cell and add a delegate to handle button clicks. Also, create a Films class to hold the data and populate the CollectionView with it. This commit sets up the basic functionality of the app.